Moved geometry.h to source/core folder
