+++ source/core/geometry.h
